test: split PDF inline source semantics cases
